+++ app/src/main/java/com/cmtech/android/bledevice/hrm/model/OnHrmListener.java
@@ -19,7 +19,7 @@ public interface OnHrmListener {
     void onHRStatisticInfoUpdated(BleHrRecord10 record);
     void onHRSensLocUpdated(int loc); // sensor location updated
     void onHRCtrlPtUpdated(int ctrl); // control point updated
-    void onFragmentUpdated(int sampleRate, int value1mV, double zeroLocation, boolean inHrMode); // fragment updated
+    void onFragmentUpdated(int sampleRate, int value1mV, float zeroLocation, boolean inHrMode); // fragment updated
     void onHrRecordStatusUpdated(boolean record);
     void onEcgSignalShowed(int ecgSignal); // ecg signal showed
     void onEcgSignalRecordStatusUpdated(boolean record); // ecg signal recorded
